pixel perfection/requirements.txt
flask
boto3
werkzeug
python-dotenv
gunicorn